+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0 / 2026-02
+* Upgraded the plugin to use Akamai Edgegrid v12
+* Upgraded the plugin to use latest Prometheus Go libraries
+* Added support for the latest Akamai Reporting API -- [authoritative-dns-traffic-by-time](https://techdocs.akamai.com/reporting/v1/reference/authoritative-dns-traffic-by-time)
+* Added the --edgedns.use-legacy-api flag to allow users to use the [Edge DNS Traffic Reporting API v1](https://techdocs.akamai.com/developer/pdfs/edge-dns-traffic-reporting-api-v1.pdf)
+* Added support for darwin/arm64
+
+
 ## 0.1.0 / 2021-01-31
 
 Initial release.